Table CRUD

+++ RestaurantProject/DataStore/TableRepository.ts
@@ -1,13 +1,10 @@
 import {Schema, model, connect} from 'mongoose';
-import Table from '../CoreBusiness/TableModel';
-import Reservation from '../CoreBusiness/ReservationModel';
-import {ReservationRepository} from './ReservationRepository';
+import Table from '../CoreBusiness/Table';
 
 export class TableRepository
 {
     tableSchema = new Schema<Table>(
         {
-            id: {type: Schema.Types.ObjectId, required: false},
             number: {type: Number, required: true},
             seats: {type: Number, required: true},
             status: {type: Number, required: true}
@@ -57,9 +54,13 @@ export class TableRepository
         }
     }  
     
-    async addTable(table: Table) : Promise<void>
+    async addTable(table: Table) : Promise<boolean>
     {
         await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
+
+        const alreadyExists = await this.TableModel.findOne({number: table.number});
+        if(alreadyExists)
+            return false;
 
         await this.TableModel
         .create(table)
@@ -71,12 +72,22 @@ export class TableRepository
         {
             console.log(err);
         });
+
+        const existsAfter = await this.TableModel.findOne({number: table.number});
+        if (existsAfter)
+            return true;
+        else
+            return false;
     }
 
-    async deleteTableByNumber(tableNumber: number) : Promise<void>
+    async deleteTableByNumber(tableNumber: number) : Promise<boolean>
     {
         await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
     
+        const exists = await this.TableModel.exists({number: tableNumber});
+        if (!exists)
+            return false;
+
         await this.TableModel
         .deleteOne({number: tableNumber})
         .then(function()
@@ -86,43 +97,41 @@ export class TableRepository
         {
             console.log(err);
         });
+
+        const existsAfter = await this.TableModel.exists({number: tableNumber});
+        if (existsAfter)
+            return true;
+        else
+            return false;
     }
 
-    async getTableByNumber(tableNumber: number) : Promise<Table>
+    async getTableByNumber(tableNumber: number) : Promise<Table | boolean>
     {
         await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
     
-        let table = await this.TableModel.findOne({number: tableNumber});
+        const table = await this.TableModel.findOne({number: tableNumber});
         if (table)
             return table;
         else
-            return null as any;
+            return false;
     }
 
-    async getTableById(tableId: string) : Promise<Table>
+    async getTables() : Promise<Table[] | boolean>
     {
         await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
     
-        let table = await this.TableModel.findById(tableId);
-        if (table)
-            return table;
+        const tables = await this.TableModel.find({});
+        if (tables)
+            return tables;
         else
-            return null as any;
+            return false;
     }
 
-    async getTables() : Promise<Table[]>
-    {
-        await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
-    
-        return await this.TableModel.find({});
-    }
-
-    async updateTableByNumber(tableNumber:number, table: Table) : Promise<void>
+    async updateTableByNumber(tableNumber:number, table: Table) : Promise<boolean>
     {
         await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
 
         let tableToUpdate = await this.TableModel.findOne({number: tableNumber});
-
         if (tableToUpdate)
         {
             if(table.number)
@@ -132,7 +141,8 @@ export class TableRepository
             if(table.status)
                 tableToUpdate.status = table.status;
 
-            await tableToUpdate.save()
+            await this.TableModel
+            .updateOne({number: tableNumber}, tableToUpdate)
             .then(function()
             {
                 console.log("Table " + tableNumber + " has been updated!");
@@ -140,38 +150,9 @@ export class TableRepository
             {
                 console.log(err);
             });
+            return true;
         }
-    }
-
-    // TODO: fix this crap
-    async getFreeTables(startDateTime : Date, endDateTime: Date, numberOfPeople: number) : Promise<string>
-    {
-        await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
-
-        let reservationRepository = new ReservationRepository();
-        let tables = await this.TableModel.find({seats: {$gte: numberOfPeople}});
-        let freeTables: Table[] = [];
-
-        for(let i = 0; i < tables.length; i++)
-        {
-            let table = tables[i];
-            let reservations = await reservationRepository.getReservationsByTableId(table._id.toString());
-
-            let isFree = true;
-
-            for(let j = 0; j < reservations.length; j++)
-            {
-                let reservation = reservations[j];
-
-                if(reservation.startDateTime < endDateTime && reservation.endDateTime > startDateTime)
-
-                    isFree = false;
-            }
-
-            if(isFree)
-                freeTables.push(table);
-        }
-
-        return "przeszło";
+        else
+            return false;
     }
 }
